requirements.txt
pyserial
git+https://github.com/Andre0512/pyserial-asyncio.git
